Adds tree method to retrieve a tax tree for a version in the package
